- version bump to v0.05! - ...and I actually remembered to update the visible version number on the webpage for once <_<;
- fixed issue where the reminder about Priests getting an extra class discipline from their "Worships the [X]" Kit also appeared as a dropdown suggestion for the Class Discipline search

- fixed some Kits showing empty Requirements instead of "None" (Bind Familiar, Brews Poisons, Channels Godmind, Charms Animals and Conjures Fiends)

- added support for search dropdowns to Arts (was formerly only on Classes) and refactored how the datalists are populated to make implementing on other tables (or adding new search fields) slightly faster/easer

- changed variable and function names in the js to be more consistent (now using snake_case for most vars and camelCase for methods, which I don't think is the standard for js, but it's still better than the previous mess of using both for both x_x)

- fixed regression where the prev/next buttons on class show info were broken (because they weren't using the new enable_navigation flag that had been added to prevent the buttons showing on Tradition.html and Role.html)

- moved some vars to misc.js rather than repeatedly declaring vars with the same name to refer to the same document elements over and over (yes global variables are bad practice, but I cba to make this project follow OO principles right now or constantly rewrite them in functions and a lot of vars were global due to being outside function scope anyway)

- minor refactor of showing + removing modal show info navigation buttons and moved to misc.js as in theory nothing there is specific to Classes and shouldn't need more than some minor changes to work on other tables...or at least the ones that just open a html file, the ones that do display via json instead can't use this as a drop-in quite yet
(main change is you have to pass in the equivalent of class_name_list, the key, the "class_" prefix used as the tr id, the table and the name of the "show[X]Info" method to use...but in time I'm hoping the showinfo method can also be made generic rather than tailored per table?)

- added the modal nav to kits to prove that worked for more than just classes
(and in the process spotted + fixed a bug with it that would have been a major headache otherwise - the class version had be written with a special case that decremented the prev/next keys by 1, as the class keys start at 1 instead of 0, and thus had to -1 to be used as the class_name_list index, but the other json files usually start at key = 0. This itself is due to how csv_to_json uses the index of the csv as the key, starting after the header, but classes.csv starts with a dummy "aaaaaa" entry from the xslx. So the actual long-term fix was to just remove the special case -1 to the key to grab the index, and modify csv_to_json.py so it uses the number of lines actually saved to the json as the key, instead of the csv index. But then we still run into the issue of using a key on the index, as we skip some classes like warrior in the class.json parsing that'd potentially induce an offset, so we ALSO change it to access the class/item name of the next html file by referencing the actual json data with the prev_key and next_key instead of using it as an array index! Phew!)
(note 2 on this: this still needs further refactoring as I think the way it checks for next/prev table row can be optimised? and I think we might be able to just use item_data instead of needing name_list? TODO in future I guess, will likely be needed to make lookup when in the middle of the powers table usable without hanging the browser? Or not, seems to be usable so far, even when picking powers near the end of the table or with search terms that muddle about the key jumps, no noticable lag yet, unlike searching)

- actually, scratch what past me said about needing some work to make nav buttons work on all the tabs, I can make showModalNavigation work for the non-html showInfo tables, just needed a slight rewrite to how the html-using showInfo methods work so they grab the name.html from the key, which now means all showInfo methods use the same key + enable_navigation args instead of some needing name (for name.html) + key + enable_navigation
- confirmed working on both html and json based showInfo methods...
- ...thus, I have added prev/next buttons to ALL current tabs/tables EXCEPT Species: Classes, Powers, Cruxes, Heritages, Kits and Arts
(species is excluded just because it doesn't actually have a html showInfo display yet)

- heckin' chonker of a commit

- TODO next, probably: finish setting up datalist dropdowns, currently only Classes and Arts have them