- v0.08 - the Feat update
- finished data entry for feats.csv, added in missing feat benefit descriptions on Martial Training, Psi Focus, Shard, Aura Shard, Blast Shard, Shield Shard and Weapon Shard groups (may come back to the csv later to add in flavour text as a column, but currently very few feats have flavour text. might also be worth in future adding an option for when a feat grants a power to display the power card for that power on the feat showInfo? Also, needs a better way of handling benefit + special text with line breaks, as those newlines had to be removed for the csv)

- added toggle on Powers tab to swap between Orcus and 4e colours for power cards (using the exact shades from their respective PDFs for At-Will, Encounter, Daily and Item powers)

- update_data.py now generates feats.json from feats.csv and is slightly better about using vars for paths (as some files have been relocated)

- added Feats tab (feature parity with the other tables: multi-column search, sorting, showInfo display, dropdown search suggestions, showInfo navigation links)

- reordered the tabs slightly to follow the Core/Advanced Options split: Classes, Powers, Feats, Cruxes, Heritages, Kits; Arts, Species

- json files now stored at "data/json/" instead of at "data/" to reduce clutter

- added special-case handling for the "Blessing of the God" feat to display the table of "Domain, God, Power" as a proper html table, as formatting when it was part of the feats.csv benefit was unreadable.
(was going to do this as a csv->json, load the json in feats.js and then assemble the table dynamically as part of that feat's html, but things get muddy if we do a lot of misc json loads and I'd have to ensure it was loaded BEFORE feats.json, which I cannot easily guarantee with the XMLHttpRequest, I think? So for now it's just a hard-coded special case)

- new experimental feature: option to go to previous (or any other) showInfo modal instead of closing.
Current example usage if on the "Blessing of the God" feat, the domain powers are all links that open showPowerInfo(), normally you'd close the power and close the modal, but chances are you probably want to go back to the feat to view the other domains + powers.
With this, you can go showFeatInfo() -> domain power -> showPowerInfo() -> close -> showFeatInfo() -> close -> actual close
(also, this works regardless of if you close via the X button or clicking off the modal, and still lets the modal just close/hide on first close for everywhere else - in future could see this being used to link to powers from other tables? Or maybe even go from the Priest class feature that grants a free 'Worships the X' Kit to the actual showKitInfo...although that'll need more work as currently we just show the html generated from the markdown)