Added the some handling for URL params
Now viewing a given power, class, discipline feat, etc will actually update the URL and going to that URL will take you to the relevant info with the show info modal dialog already open

Means the compendium is much more useful as you can link a rules element like a power directly rather than having to search for it each time

Currently includes:
- Classes
- Disciplines
- "Class Misc" info (currently just Tradition and Role)
- Powers
- Feats
- Cruxes
- Heritages
- Kits
- Prestige Paths
- Epic Paths
- Companions
- Arts
- Species

Also added params for handling Power Card style (power_colour and power_format, both with values Orcus/Compendium or 4e style), which auto-sets the appropriate radio buttons on the power card selection

Can also now hide the tables and no longer shows classes by default - gives an easier option to see the changelog + known issues without having to scroll as much.

BUGFIX: from last commit the code for displaying what sources a power comes from was broken due to the change in formatting, as the old code seems to expect every class and kit to have (class) and (kit) after them, which breaks on some powers as the code just does a split on brackets and doesn't handle cases where (class) or (kit) isn't present

Fix was to only do the split if brackets are actually there for class/kit, as if they're not there, the array element already is a valid class/kit name to do a lookup with.